+++ nft-ticket/programs/nft-ticket/src/state/ticket.rs
@@ -8,7 +8,7 @@ pub struct Ticket {
     pub taker: Pubkey,
     pub nft_mint: Option<Pubkey>,
     pub bump: u8,
-    pub max_supply: u64,
+    pub max_supply: u16,
 }
 impl Space for Ticket {
     const INIT_SPACE: usize = 8 + 32 + 2 + 8 + 32 + 32 + 1 + 8;
